Daily_Weekly_tweets/requirements.txt
pandas==0.25.0
python-dotenv==0.10.3
twython==3.7.0
bert-serving-client==1.9.6
bert-serving-server==1.9.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.0
tensorflow-estimator==1.14.0
tweepy==3.8.0
requests-oauthlib==1.2.0
selenium==3.141.0
requests==2.20.0
tensorflow-gpu==1.14.0